Add the necessary panier service methods (crud)

+++ src/main/java/ma/ac/emi/ginf/emishop/Model/PanierItem.java
@@ -29,7 +29,7 @@ public class PanierItem {
         return id;
     }
 
-    public int getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
